requirements/base.txt
###############################################
# Django и основные зависимости (ядро проекта)
###############################################
Django==5.1
djangorestframework==3.15.2
channels==4.2.2
asgiref==3.8.1
django-environ==0.12.0
django-import-export==4.3.7
django-json-widget==2.0.1
django-autocomplete-light==3.12.1
django-unfold==1.0.0
django-debug-toolbar==5.2.0

#############################
# База данных и драйверы
#############################
psycopg2-binary==2.9.10
SQLAlchemy==2.0.35
fdb==2.0.3

########################
# Стек Dash (frontend)
########################
dash==3.0.4
dash-ag-grid==31.3.1
dash-bootstrap-components==2.0.3
dash-html-components==2.0.0
dash-table==5.0.0
plotly~=6.3.0

############################
# ETL и Dagster экосистема
############################
dagster==1.10.5
dagit==1.10.5

############################
# Обработка и анализ данных
############################
pandas==2.2.3
openpyxl==3.1.5
pillow==11.2.1
beautifulsoup4==4.13.4

############################
# Интеграции и утилиты
############################
requests==2.32.3
paramiko==3.5.1
psutil==5.9.8
selenium==4.32.0
webdriver-manager==4.0.2
python-telegram-bot==22.0
python-dotenv==1.1.0
setuptools==80.7.1
chardet==5.2.0